requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.2
python-jose==3.3.0
passlib==1.7.4
